# Adding new loop_parse_if_digits function