--- a/include/fast_float/constexpr_feature_detect.h
+++ b/include/fast_float/constexpr_feature_detect.h
@@ -14,7 +14,7 @@
 #define FASTFLOAT_CONSTEXPR14
 #endif
 
-#if __cplusplus >= 201703L
+#if __cplusplus >= 201703L || (defined(_MSVC_LANG) && _MSVC_LANG >= 201703L)
 #define FASTFLOAT_IF_CONSTEXPR if constexpr 
 #else
 #define FASTFLOAT_IF_CONSTEXPR if