@lemire 2024-11-25
PR https://github.com/fastfloat/fast_float/pull/293 added benchmarks to the project. I invite you to sync your fork and run these benchmarks (before and after), ideally with performance counters.