@RealTimeChris 2024-11-25
> PR #293 added benchmarks to the project. I invite you to sync your fork and run these benchmarks (before and after), ideally with performance counters.

Alright thank you very much I will do that.