@RealTimeChris 2024-11-25
> I encourage you to include a rationale as to why you think that compilers will generate faster code.
> 
> Be mindful that if you are recording timings on purely synthetic benchmarks, without performance counters, you can easily derive the wrong conclusions.
> 
> The bar is pretty high for us to accept an optimization in this library as it is quite mature. It is fairly easy to reorganize the code and get better results under on compiler for one type of synthetic data set.
> 
> Please consider there is a significant cost for us to accepting such as change as it may introduce a new bug, which would be terrible. So we need hard evidence that the result is definitively faster.
> 
> For benchmarking, I use the following framework...
> 
> https://github.com/lemire/simple_fastfloat_benchmark
> 
> It includes realistic files (derived from actual content).
> 
> I tried testing your PR, but it won't build.
> 
> We do very much encourage pull requests which might improve performance, but, as I wrote, the bar is high. I will be definitively doing my own analysis, on different hardware and different compilers.

I included that. As I said, given that we are no longer subtracting 0x30303030303030 twice, here:
https://github.com/fastfloat/fast_float/blob/d4c573de0b5a1c934f719e6a929396746d3849e6/include/fast_float/ascii_number.h#L143
and here:
https://github.com/fastfloat/fast_float/blob/d4c573de0b5a1c934f719e6a929396746d3849e6/include/fast_float/ascii_number.h#L124
And instead moving the subtraction out to here:
https://github.com/fastfloat/fast_float/blob/d4c573de0b5a1c934f719e6a929396746d3849e6/include/fast_float/ascii_number.h#L244
As well as no longer collecting the value twice as is done here:
https://github.com/fastfloat/fast_float/blob/d4c573de0b5a1c934f719e6a929396746d3849e6/include/fast_float/ascii_number.h#L227-L230
But instead only collect it once here:
https://github.com/fastfloat/fast_float/blob/d4c573de0b5a1c934f719e6a929396746d3849e6/include/fast_float/ascii_number.h#L244
This should produce faster code, no?