@lemire 2024-11-25
> This should produce faster code, no?

Your mental model is that you are reducing the number of instructions. If so, then the next step is to validate that it is true. If you run the benchmark I have just included in privileged mode (so that you have access to hardware performance counters), you should see the number of instructions retired.

E.g., the output should look like this...

```
$ ./build/benchmarks/realbenchmark
...
fastfloat (64)                          :  1164.85 MB/s (+/- 1.8 %)    66.94 Mfloat/s      18.32 i/B ...
```

See the `18.32 i/B`? This is the number of instructions (on average) per byte of input. If this number goes down sufficiently, then, everything else being equal, the performance might be improved.

This depends crucially on the compiler and on the targeted system.


Alternatively, you can examine the binary output and manually check that there are fewer instructions being executed.

There are other elements to consider... but if you can show a substantial reduction in the number of retired instructions per input byte, then it is a very good start. (This is not the only way to optimize the code.)